Fix sanityCheck failures
