+++ subprojects/code-quality/src/main/groovy/org/gradle/api/plugins/quality/internal/CheckstyleAntInvoker.groovy
@@ -20,7 +20,6 @@ import org.gradle.api.GradleException
 import org.gradle.api.InvalidUserDataException
 import org.gradle.api.file.FileCollection
 import org.gradle.api.plugins.quality.Checkstyle
-import org.gradle.api.plugins.quality.CheckstyleActionParameters
 import org.gradle.internal.logging.ConsoleRenderer
 import org.gradle.util.internal.GFileUtils
 import org.slf4j.Logger
